pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.2",
    "ddgs>=9.0.0",
    "faiss-cpu>=1.8.0",
    "google-generativeai>=0.5.2",
    "ipykernel>=7.1.0",
    "langchain>=0.2.1",
    "langchain-community>=0.2.1",
    "langchain-google-genai>=1.0.6",
    "langgraph>=0.0.45",
    "pydantic>=2.5.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "streamlit>=1.31.0",
    "wikipedia>=1.4.0",
]
